% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-ci.R
\name{BinomCIMeanBootCustom}
\alias{BinomCIMeanBootCustom}
\title{Calculate confidence interval for the mean of binomial proportions using bootstrapping}
\usage{
BinomCIMeanBootCustom(x, n, alpha = 0.05, B = 1000)
}
\arguments{
\item{x}{matrix or numeric. First column of the matrix, or the numeric vector, should represent the estimated binomial proportion for each experiment.}

\item{n}{integer. Number of trials per experiment. If \code{length(n)} does not match the number of estimated binomial proportions it is repeated.}

\item{alpha}{numeric. Probability level for Z-value in normal distribution confidence interval calculation.}

\item{B}{integer. Number of bootstrap replicates to use. Default: \code{1000}}
}
\value{
matrix containing columns \code{"est"}, \code{"lwr.ci"}, \code{"upr.ci"}
}
\description{
This method accounts for unique confidence intervals from each experiment by sampling from the confidence intervals during bootstrapping. Upper and lower confidence interval boundaries are determined using quantiles (corresponding to the specified level of \code{alpha}) of the bootstrapped estimates.
}
\examples{

library(DescTools)

# example: 3 binomial experiments, 10 trials, confidence level 0.8
x <- DescTools::BinomCI(c(5,6,7), 10, 0.8)

# Inspect input
x

# Calculate mean binomial probability and CI using bootstrapping
# NOTE: the estimated CI from BinomCI() are considered in this method
BinomCIMeanBootCustom(x, 10, B = 100)

# Compare to normal approximation
BinomCIMeanNorm(x, 10)

# Compare to straight arithmetic average of experiments
colMeans(x)

}
