% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-ci.R
\name{BinomCIMeanPoolCustom}
\alias{BinomCIMeanPoolCustom}
\title{Calculate confidence interval for the mean of binomial probabilities using pooled variance}
\usage{
BinomCIMeanPoolCustom(x, n, alpha = 0.05)
}
\arguments{
\item{x}{matrix or numeric. First column of the matrix, or the numeric vector, should represent the estimated binomial proportion for each experiment.}

\item{n}{integer. Number of trials per experiment. If \code{length(n)} does not match the number of estimated binomial proportions it is repeated.}

\item{alpha}{numeric. Probability level for Z-value in normal distribution confidence interval calculation.}
}
\value{
matrix containing columns \code{"est"}, \code{"lwr.ci"}, \code{"upr.ci"}
}
\description{
This method accounts for unique confidence intervals from each experiment by calculating the pooled variance adjusted for the confidence interval around each binomial probability. It does not use bootstrapping or rely on a large number of trials per experiment (though, the number of trials should not be \emph{very} small; i.e \code{n} should be at least 5, ideally larger) It does not necessarily rely on many repetitions of each experiment because the final mean binomial probability is calculated using the t-distribution and number of degrees of freedom corresponding to the effective sample size. The effective sample size is used to calculate a standard error from the pooled variance across experiments and construct the final confidence interval.
}
\examples{

library(DescTools)

# example: 3 binomial experiments, 10 trials, confidence level 0.8
x <- DescTools::BinomCI(c(5,6,7), 10, 0.8)

# Inspect input
x

# Calculate confidence interval for mean binomial probability
# NOTE: accounting for custom CI in pooled variance
BinomCIMeanPoolCustom(x, 10)

# Compare to t-distribution
BinomCIMeanPoolNorm(x, 10)

# Compare to arithmetic average
colMeans(x)

}
