% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-ci.R
\name{BinomCIMeanPoolNorm}
\alias{BinomCIMeanPoolNorm}
\title{Calculate confidence interval for the mean of binomial probabilities using pooled variance}
\usage{
BinomCIMeanPoolNorm(x, n, alpha = 0.05)
}
\arguments{
\item{x}{matrix or numeric. First column of the matrix, or the numeric vector, should represent the estimated binomial proportion for each experiment.}

\item{n}{integer. Number of trials per experiment. If \code{length(n)} does not match the number of estimated binomial proportions it is repeated.}

\item{alpha}{numeric. Probability level for Z-value in normal distribution confidence interval calculation.}
}
\value{
matrix containing columns \code{"est"}, \code{"lwr.ci"}, \code{"upr.ci"}
}
\description{
This method estimates confidence intervals for each unique experiment using the normal approximation. This is suitable for when the number of trials in each experiment is large enough such that Central Limit Theorem holds (i.e. 30+ trials per experiment). It does not necessarily rely on many repetitions of each experiment because the final mean binomial probability is calculated using the t-distribution and number of degrees of freedom corresponding to the effective sample size. The effective sample size is used to calculate a standard error from the pooled variance across experiments and construct the final confidence interval.
}
\examples{

library(DescTools)

# example: 3 binomial experiments, 10 trials, confidence level 0.8
x <- DescTools::BinomCI(c(5,6,7), 10, 0.8)

# Inspect input
x

BinomCIMeanPoolNorm(x, 10)

# Compare to arithmetic average
colMeans(x)

}
