% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled-variance.R
\name{BinomVarCustom}
\alias{BinomVarCustom}
\title{Calculate pooled variance for related binomial experiments with given confidence intervals}
\usage{
BinomVarCustom(n, p_values, ci_intervals)
}
\arguments{
\item{n}{integer. Number of trials per experiment}

\item{p_values}{numeric. Probability of success for each experiment.}

\item{ci_intervals}{list of numeric vectors. The list has length equal to length of \code{p_values}. Each element of the list contains a numeric vector of length 2 denoting the lower and upper confidence limits.}
}
\description{
This method calculates pooled variance for related binomial experiments with confidence intervals. Instead of relying on the normal approximation, the user can calculate or estimate their own confidence intervals. The method adjusts the calculated binomial variance to account for an arbitrary given confidence interval, and then calculates pooled variance across all experiments.
}
\examples{

n <- 10  # Number of trials for each experiment

p_values <- c(0.2, 0.3, 0.4) # Probability of success for each experiment

# Wilson score intervals for each p
wilson_intervals <- list(
  c(0.1, 0.3),
  c(0.2, 0.4),
  c(0.3, 0.5)
)

# Calculate pooled variance with custom confidence intervals
BinomVarCustom(n, p_values, wilson_intervals)
}
