% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{greekletters}
\alias{greekletters}
\title{Table of Greek Letters in upper and lowercase, and ASCII equivalent}
\format{
An object of class \code{data.frame} with 24 rows and 4 columns.
}
\usage{
data(greekletters)
}
\description{
A data.frame containing the name, upper case, and lowercase (unicode) characters, along with ASCII equivalents.
}
\keyword{datasets}
