% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{loamy}
\alias{loamy}
\title{Synthetic Site-level Properties for Loamy Soils}
\format{
An object of class \code{data.frame} with 30 rows and 5 columns.
}
\usage{
data(loamy, package = "SOILmilaR")
}
\description{
A \emph{data.frame} containing a unique profile ID, depth, particle
size class, along with particle size control section depth-weighted average
clay content and rock fragment volume.
}
\keyword{datasets}
