% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscs_weighted.R
\name{pscs_weighted}
\alias{pscs_weighted}
\title{Calculate Particle Size Control Section Weighted Average}
\usage{
pscs_weighted(
  x,
  vars,
  na.rm = FALSE,
  pscsbounds = c("psctopdepth", "pscbotdepth"),
  prefix = "pscs_"
)
}
\arguments{
\item{x}{A \emph{SoilProfileCollection}}

\item{vars}{\emph{character}. Column names of numeric variables in horizon table of \code{x} to summarize.}

\item{na.rm}{\emph{logical}. Remove \code{NA}? Default: \code{FALSE}}

\item{pscsbounds}{\emph{character} of length 2. Column names of numeric variables in site table representing the profile specific upper and lower boundary. Default: \code{"psctopdepth"} and \code{"pscbotdepth"}}

\item{prefix}{\emph{character}. Prefix to append before column name, to create unique name that does not conflict with existing horizon data. Default: \code{"pscs_"}}
}
\value{
A data.frame-like object corresponding to \code{aqp_df_class(x)}
}
\description{
Calculates the particle size control section weighted average for one or more variables (\code{vars}), using upper and lower boundary specified in \code{pscsbounds}. Profiles are truncated to the interval set by these bounds, and weighted averages are calculated for each column. The result has the same number of rows as the number of sites in the input SoilProfileCollection \code{x} and is suitable for joining back into the site table
}
\examples{
\dontshow{if (!inherits(try(requireNamespace("aqp", quietly = TRUE) && requireNamespace("soilDB", quietly = TRUE), silent = TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# load aqp, and data from soilDB
library(aqp)
data(loafercreek, package = "soilDB")

# calculate PSCS weighted clay \% and total fragment volume
pscs_cf <- pscs_weighted(loafercreek, c("clay", "fragvoltot"))

# inspect
pscs_cf

# left join to site table of loafercreek
site(loafercreek) <- pscs_cf
\dontshow{\}) # examplesIf}
}
