% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOILmilaR-package.R
\docType{package}
\name{SOILmilaR-package}
\alias{SOILmilaR}
\alias{SOILmilaR-package}
\title{SOILmilaR: Apply Similar Soils Contrasts}
\description{
Provides methods for applying standardized, customizable "similar soils" rules to site-level data derived from various sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/brownag/SOILmilaR}
  \item \url{http://humus.rocks/SOILmilaR/}
}

}
\keyword{internal}
