% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_soils.R
\name{similar_soils}
\alias{similar_soils}
\title{Similar Soil Contrasts}
\usage{
similar_soils(
  x,
  mapping,
  condition = NULL,
  idname = "id",
  thresh = NULL,
  thresh_single = 2,
  thresh_all = 3,
  absolute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \emph{data.frame} or a \emph{SoilProfileCollection}}

\item{mapping}{A named \emph{list} of functions. List element names refer to
site-level data columns of \code{x}. Each specified function provides a
conversion of the data element in \code{x} to a value used in the "similar
soils" calculation.}

\item{condition}{integer or character. Default: \code{NULL} the value is
calculated internally based on the dominant condition of intersection of
mapping results in \code{x}. You may specify an integer row ID in \code{x} for
specific similar soil contrasts, or you can specify a character dominant
condition equivalent to the name assigned by \code{interaction()} e.g. \code{"4.3"} for
a two rating mapping result where the first rating has value \code{4} and the
second has value \code{3}}

\item{idname}{ID column name, default \code{"id"}}

\item{thresh}{Deprecated. If used passed to \code{thresh_all}.}

\item{thresh_single}{Sum of differences relative to dominant condition in
\code{x}. Default \code{1}. See details for discussion on the default calculation for
similarity.}

\item{thresh_all}{Sum of differences relative to dominant condition in \code{x}.
Default \code{1}. See details for discussion on the default calculation for
similarity.}

\item{absolute}{logical. Report absolute difference? Default: \code{TRUE}.
Absolute difference is always used for comparison against thresh.}

\item{verbose}{Default: \code{TRUE} message about selected \code{condition}}
}
\value{
A \emph{data.frame} containing inputs and three new columns:
\code{similar_single} (maximum difference in any one property, relative to
\code{condition}), \code{similar_dist} (cumulative sum of differences relative to
\code{condition}), \code{similar} (logical; soil is similar to \code{condition})
}
\description{
A method for applying standardized, customizable "similar soils" rules to
site-level data derived from various sources.
}
\details{
The sum of differences across conditions (specified by the intersection of
output of the functions in \code{mapping}) is used as the "distance" of a soil
relative to a dominant (or otherwise specified) condition. A threshold value
is used to decide which are "similar" and which are not. The functions in
mapping can be customized to use alternate thresholds.
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly=TRUE) && requireNamespace("cluster", quietly=TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

data(loamy, package = "SOILmilaR")

rate_taxpartsize <- function(x) {
  dplyr::case_match(x,
                    c("sandy-skeletal") ~ 1,
                    c("sandy") ~ 3,
                    c("loamy", "coarse-loamy", "coarse-silty") ~ 5,
                    c("fine-loamy", "fine-silty") ~ 7,
                    c("clayey", "fine") ~ 9,
                    c("very-fine") ~ 11,
                    c("loamy-skeletal", "clayey-skeletal") ~ 13,
                    "fragmental" ~ 15)
}


rate_depthclass <- function(x, breaks = c( `very shallow` = 25, `shallow` =
  50, `moderately deep` = 100, `deep` = 150, `very deep` = 1e4 ), pattern =
  "R|Cr|Cd|kk|m", hzdesgn = aqp::guessHzDesgnName(x, required = TRUE),
                            ...) {
  res <- cut(x, c(0, breaks))
  factor(res, levels = levels(res), labels = names(breaks), ordered = TRUE)
}

rate_pscs_clay <- function(x,
                           breaks = c(18, 27, 40, 60, 100)) {
  res <- cut(x, c(0, breaks))
  factor(res, levels = levels(res), ordered = TRUE)
}

m <- list(taxpartsize = rate_taxpartsize, depth = rate_depthclass,
pscs_clay = rate_pscs_clay)

s <- similar_soils(loamy, m)
head(s)

# inspect distances using agglomerative clustering+dendrogram
d <- cluster::agnes(s[, 5, drop = FALSE], method="gaverage")
d$height <- d$height + 0.2 # fudge factor for 0-distance
plot(stats::as.dendrogram(d), center=TRUE, type="triangle")

# allow relative contrast ratings to be negative # (i.e. ordinal factors, concept of "limiting")
# absolute value is still used for "similar" threshold
s2 <- similar_soils(loamy, m, absolute=FALSE)

# inspect distances using agglomerative clustering+dendrogram
d2 <- cluster::agnes(s2[, 5, drop = FALSE], method="gaverage")
d2$height <- d2$height + 0.2 # fudge factor for 0-distance
plot(stats::as.dendrogram(d2), center=TRUE, type="triangle")
\dontshow{\}) # examplesIf}
}
\references{
Norfleet, M.L. and Eppinette, R.T. (1993), A Mathematical Model
for Determining Similar and Contrasting Inclusions for Map Unit
Descriptons. Soil Survey Horizons, 34: 4-5.
\url{https://doi.org/10.2136/sh1993.1.0004}
}
